requirements.txt
Django==1.11.9
djangorestframework==3.7.7
django-rest-swagger==2.1.2
django-filter==1.1.0
behave-django==1.0.0
django-allauth==0.34.0
django-rest-auth==0.9.3
djangorestframework-jwt==1.11.0
pytz==2017.3
